latinIME: fix NumberFormatException
tbh I do not know what google at all thought about doing this approach
on getting the AutoCorrectionThreshold. At least the whole method can
be done in a way cleaner way. Currently due that on config.xml the
corespondent entry is empty when autocorrect mode is turned off the method
throws every time a NumberFormatException.

Whatever for first let us catch the case that the value is empty
and set it to Float.MAX_VALUE to avoid this issue

will revisit it on a later point and rewrite some parts on LatinIME
due that this weird approach is not only used here

Change-Id: Ifbd1494c9ef30265545e1eab4305d0d671e45fd1